bot_food/requirements.txt
python==3.9.13 
python-telegram-bot==13.15
python-dotenv==1.1.0